requirements.txt
pandas==1.5.3
numpy==1.26.4
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.2
python-dateutil>=2.9
pytz>=2024.1
tenacity>=8.2
tqdm>=4.66
PyYAML>=6.0
rapidfuzz>=3.9
nfl-data-py==0.3.3
